modificações para a preparação do roteiro
