--- src/br/com/alura/kotlin/collection/list/PrateleiraMutavel.kt
@@ -1,6 +1,0 @@
-package br.com.alura.kotlin.collection.list
-
-data class PrateleiraMutavel (
-    val genero: String,
-    val livros: MutableList<Livro>
-)
